+++ CHANGELOG.md
@@ -24,6 +24,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - App: Rename ´definitions´ to ´resources´ in app definition.
 - Block: Rewrite and simplify detail-viewer block.
 - Block: Replace usage of resource in map block with a load action.
+- Server: Add security scopes for API calls.
 
 ## [0.4.2] - 2019-02-26